+++ e2e/tests/assorted/05-joinpublicroom.spec.js
@@ -15,7 +15,7 @@ async function navigateToRoom() {
 }
 
 async function navigateToRoomActions() {
-	await element(by.id('room-view-header-actions')).tap();
+	await element(by.id('room-header')).tap();
 	await waitFor(element(by.id('room-actions-view'))).toBeVisible().withTimeout(5000);
 }
 
@@ -39,7 +39,7 @@ describe('Join public room', () => {
 		// Render - Header
 		describe('Header', async() => {
 			it('should have actions button ', async() => {
-				await expect(element(by.id('room-view-header-actions'))).toBeVisible();
+				await expect(element(by.id('room-header'))).toBeVisible();
 			});
 		});
